requirements.txt
azure-eventhub
influxdb